corrige salvamento de quantidade/nivelCritico e exibição de estoque na bomboniere

+++ presentation-backend/src/main/java/br/com/cinema/frame/bomboniere/InsumoRequest.java
@@ -3,7 +3,7 @@ package br.com.cinema.frame.bomboniere;
 public record InsumoRequest(
         String nome,
         String unidade, 
-        Integer quantidade,
-        Integer nivelCritico
+        Double quantidade,
+        Double nivelCritico
 ) {
 }
